Small changes & fixes
Pages folder moved into the components folder.
Final calendar changes.
Small changes everywhere else.
