+++ frontend/src/App.js
@@ -4,22 +4,22 @@ import React from 'react';
 import useTokenExpirationChecker from './hooks/useTokenExpirationChecker';
 
 // pages & components
-import Home from './pages/Home'
+import Home from './components/pages/Home'
 import Navbar from './components/Navbar'
 import Bottombar from './components/Bottombar'
-import TaskForm from './pages/TaskForm'
-import CalendarPage from './pages/Calendar'
-import Overview from './pages/Overview'
-import History from './pages/History'
-import EditHistory from './pages/EditHistory'
-import Profile from './pages/Profile'
-import LoggedOut from './pages/LogOut'
-import Login from './pages/Login'
-import Signup from './pages/Signup'
-import Verify from './pages/Verification'
-import ForgotPassword from './pages/ForgotPassword';
-import ResetPassword from './pages/ResetPassword';
-import Users from './pages/Users';
+import TaskForm from './components/pages/TaskForm'
+import CalendarPage from './components/pages/Calendar'
+import Overview from './components/pages/Overview'
+import History from './components/pages/History'
+import EditHistory from './components/pages/EditHistory'
+import Profile from './components/pages/Profile'
+import LoggedOut from './components/pages/LogOut'
+import Login from './components/pages/Login'
+import Signup from './components/pages/Signup'
+import Verify from './components/pages/Verification'
+import ForgotPassword from './components/pages/ForgotPassword';
+import ResetPassword from './components/pages/ResetPassword';
+import Users from './components/pages/Users';
 
 
 function App() {
@@ -32,7 +32,7 @@ function App() {
   const {user} = useAuthContext()
   
   return (
-    <div>
+    <div className='root'>
       <BrowserRouter>
         <Navbar />
         <TokenExpirationChecker />
@@ -99,7 +99,7 @@ function App() {
               element={user && user.role !== 'employee'? <Users /> : <Navigate to="/login" />}
               />
             </Routes>
-        </div>
+          </div>
           <Bottombar />
           </BrowserRouter>
     </div>
