+++ frontend/src/components/pages/EditHistory.js
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from 'react';
-import { useTasksContext } from '../hooks/useTasksContext';
-import { useCustomFetch } from '../hooks/useCustomFetch'; // Make sure to import useCustomFetch
+import { useTasksContext } from '../../hooks/useTasksContext';
+import { useCustomFetch } from '../../hooks/useCustomFetch'; // Make sure to import useCustomFetch
 import { useNavigate, useParams } from 'react-router-dom';
-import './../css/TaskForm.css';
-import { useAuthContext } from '../hooks/useAuthContext';
-import { useLogout } from '../hooks/useLogout'; // Import useLogout for handling token expiration
+import '../../css/TaskForm.css';
+import { useAuthContext } from '../../hooks/useAuthContext';
+import { useLogout } from '../../hooks/useLogout'; // Import useLogout for handling token expiration
 import moment from 'moment-timezone';
