+++ frontend/src/components/pages/Login.js
@@ -1,8 +1,8 @@
 import { useState } from 'react';
-import { useLogin } from '../hooks/useLogin';
+import { useLogin } from '../../hooks/useLogin';
 import { Link } from 'react-router-dom';
 import { useNavigate } from 'react-router-dom';
-import './../css/Login.css';
+import '../../css/Login.css';
 
 const Login = () => {
     const [email, setEmail] = useState('');
