+++ frontend/src/css/Bottombar.css
@@ -3,8 +3,8 @@
   position: fixed;
   bottom: 0;
   left: 0;
-  width: 100%;
-  height: 4%;
+  width: 100vw;
+  height: 3vh;
   z-index: 1000;
   background-image: url('https://res.cloudinary.com/djuqkjsl4/image/upload/v1698908449/CN_dchz50.png');
   background-repeat: no-repeat;
