+++ frontend/src/css/Calendar.css
@@ -1,15 +1,9 @@
 .calendar-page {
   display: flex;
-  border-radius: 4px;
-  background: #b9b9b9;
+  background: #353535;
+  margin: 1px 0;
 }
 
-.main-container {
-  display: flex;
-  background: #353535;
-  border-radius: 4px;
-  width: 100%;
-}
 
 .rbc-calendar {
   background: #353535;
@@ -18,7 +12,7 @@
   margin-left: 0;
   padding-left: 0;
   width: stretch;
-  height: 100vh; 
+  height: 85vh; 
 }
 
 /*Tool Bar*/
@@ -211,6 +205,7 @@
   width: 100%;
   background: #353535;
   padding-bottom: 2em;
+  overflow-y: auto;
 }
 
 /*Pop-up*/
@@ -282,6 +277,10 @@
     opacity: 1;
     color: black;
   }
+
+  .task-info-c .selectedEmployee {
+    font-size: smaller;
+  }
   
 .task-info-c textarea {
   padding: 0 2%;
@@ -312,14 +311,15 @@
   background: #353535;
   border: 1px solid #fff;
   color: #fff;
-  width: 23%;
+  width: fit-content;
   border-radius: 4px;
-  font-size: medium;
+  font-size: small;
   font-weight: 600;
   cursor: pointer;
   margin: 1.5%;
   margin-left: auto;
   margin-right: auto;
+  padding: 0.3rem;
   display: block;
   transition: background-color 0.3s;
 }
