+++ frontend/src/css/Navbar.css
@@ -1,12 +1,13 @@
 /* Navbar.css */
 /* Main navbar container */
 .navbar {
+  height: 5vh;
   background: #353535;
   color: #fff;
   display: flex;
   justify-content: flex-start; /* Align items to the left */
   align-items: center;
-  padding: 10px;
+  padding: 10px 20px;
   box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
   position: relative;
 }
